--- a/core/include/detray/geometry/shapes/cuboid3D.hpp
+++ b/core/include/detray/geometry/shapes/cuboid3D.hpp
@@ -62,15 +62,15 @@ class cuboid3D {
     template <template <typename, std::size_t> class bounds_t,
               typename scalar_t, std::size_t kDIM, typename point_t,
               typename std::enable_if_t<kDIM == e_size, bool> = true>
-    DETRAY_HOST_DEVICE inline bool check_boundaries(
+    DETRAY_HOST_DEVICE inline auto check_boundaries(
         const bounds_t<scalar_t, kDIM> &bounds, const point_t &loc_p,
         const scalar_t tol = std::numeric_limits<scalar_t>::epsilon()) const {
-        return (bounds[e_min_x] - tol <= loc_p[0] and
-                bounds[e_min_y] - tol <= loc_p[1] and
-                bounds[e_min_x] - tol <= loc_p[2] and
-                loc_p[0] <= bounds[e_max_x] + tol and
-                loc_p[1] <= bounds[e_max_y] + tol and
-                loc_p[2] <= bounds[e_max_z] + tol);
+        return ((bounds[e_min_x] - tol) <= loc_p[0] &&
+                (bounds[e_min_y] - tol) <= loc_p[1] &&
+                (bounds[e_min_x] - tol) <= loc_p[2] &&
+                loc_p[0] <= (bounds[e_max_x] + tol) &&
+                loc_p[1] <= (bounds[e_max_y] + tol) &&
+                loc_p[2] <= (bounds[e_max_z] + tol));
     }
 
     /// @brief Measure of the shape: Volume