--- a/core/include/detray/geometry/shapes/cylinder2D.hpp
+++ b/core/include/detray/geometry/shapes/cylinder2D.hpp
@@ -62,12 +62,12 @@ class cylinder2D {
     template <template <typename, std::size_t> class bounds_t,
               typename scalar_t, std::size_t kDIM, typename point_t,
               typename std::enable_if_t<kDIM == e_size, bool> = true>
-    DETRAY_HOST_DEVICE inline bool check_boundaries(
+    DETRAY_HOST_DEVICE inline auto check_boundaries(
         const bounds_t<scalar_t, kDIM> &bounds, const point_t &loc_p,
         const scalar_t tol = std::numeric_limits<scalar_t>::epsilon()) const {
 
-        return (bounds[e_n_half_z] - tol <= loc_p[1] and
-                loc_p[1] <= bounds[e_p_half_z] + tol);
+        return ((bounds[e_n_half_z] - tol) <= loc_p[1] &&
+                loc_p[1] <= (bounds[e_p_half_z] + tol));
     }
 
     /// @brief Measure of the shape: Area