--- a/core/include/detray/geometry/shapes/line.hpp
+++ b/core/include/detray/geometry/shapes/line.hpp
@@ -72,7 +72,7 @@ class line {
     template <template <typename, std::size_t> class bounds_t,
               typename scalar_t, std::size_t kDIM, typename point_t,
               typename std::enable_if_t<kDIM == 2u, bool> = true>
-    DETRAY_HOST_DEVICE inline bool check_boundaries(
+    DETRAY_HOST_DEVICE inline auto check_boundaries(
         const bounds_t<scalar_t, kDIM> &bounds, const point_t &loc_p,
         const scalar_t tol = std::numeric_limits<scalar_t>::epsilon()) const {
 
@@ -82,9 +82,9 @@ class line {
         // line from the line center is less than the half line length
         if constexpr (square_cross_sect) {
             return (math::fabs(loc_p[0] * math::cos(loc_p[2])) <=
-                        bounds[e_cross_section] + tol &&
+                        (bounds[e_cross_section] + tol) &&
                     math::fabs(loc_p[0] * math::sin(loc_p[2])) <=
-                        bounds[e_cross_section] + tol &&
+                        (bounds[e_cross_section] + tol) &&
                     math::fabs(loc_p[1]) <= bounds[e_half_z] + tol);
 
         }
@@ -93,8 +93,8 @@ class line {
         // of closest approach on the line from the line center is less than the
         // line half length
         else {
-            return (math::fabs(loc_p[0]) <= bounds[e_cross_section] + tol &&
-                    math::fabs(loc_p[1]) <= bounds[e_half_z] + tol);
+            return (math::fabs(loc_p[0]) <= (bounds[e_cross_section] + tol) &&
+                    math::fabs(loc_p[1]) <= (bounds[e_half_z] + tol));
         }
     }