--- a/core/include/detray/geometry/shapes/rectangle2D.hpp
+++ b/core/include/detray/geometry/shapes/rectangle2D.hpp
@@ -55,11 +55,11 @@ class rectangle2D {
     template <template <typename, std::size_t> class bounds_t,
               typename scalar_t, std::size_t kDIM, typename point_t,
               typename std::enable_if_t<kDIM == e_size, bool> = true>
-    DETRAY_HOST_DEVICE inline bool check_boundaries(
+    DETRAY_HOST_DEVICE inline auto check_boundaries(
         const bounds_t<scalar_t, kDIM> &bounds, const point_t &loc_p,
         const scalar_t tol = std::numeric_limits<scalar_t>::epsilon()) const {
-        return (math::fabs(loc_p[0]) <= bounds[e_half_x] + tol and
-                math::fabs(loc_p[1]) <= bounds[e_half_y] + tol);
+        return (math::fabs(loc_p[0]) <= (bounds[e_half_x] + tol) &&
+                math::fabs(loc_p[1]) <= (bounds[e_half_y] + tol));
     }
 
     /// @brief Measure of the shape: Area