--- a/core/include/detray/geometry/shapes/ring2D.hpp
+++ b/core/include/detray/geometry/shapes/ring2D.hpp
@@ -58,12 +58,12 @@ class ring2D {
     template <template <typename, std::size_t> class bounds_t,
               typename scalar_t, std::size_t kDIM, typename point_t,
               typename std::enable_if_t<kDIM == e_size, bool> = true>
-    DETRAY_HOST_DEVICE inline bool check_boundaries(
+    DETRAY_HOST_DEVICE inline auto check_boundaries(
         const bounds_t<scalar_t, kDIM> &bounds, const point_t &loc_p,
         const scalar_t tol = std::numeric_limits<scalar_t>::epsilon()) const {
 
-        return (loc_p[0] + tol >= bounds[e_inner_r] and
-                loc_p[0] <= bounds[e_outer_r] + tol);
+        return ((loc_p[0] + tol) >= bounds[e_inner_r] &&
+                loc_p[0] <= (bounds[e_outer_r] + tol));
     }
 
     /// @brief Measure of the shape: Area