--- a/core/include/detray/geometry/shapes/trapezoid2D.hpp
+++ b/core/include/detray/geometry/shapes/trapezoid2D.hpp
@@ -61,16 +61,16 @@ class trapezoid2D {
     template <template <typename, std::size_t> class bounds_t,
               typename scalar_t, std::size_t kDIM, typename point_t,
               typename std::enable_if_t<kDIM == e_size, bool> = true>
-    DETRAY_HOST_DEVICE inline bool check_boundaries(
+    DETRAY_HOST_DEVICE inline auto check_boundaries(
         const bounds_t<scalar_t, kDIM> &bounds, const point_t &loc_p,
         const scalar_t tol = std::numeric_limits<scalar_t>::epsilon()) const {
-        const scalar_t rel_y{(bounds[e_half_length_2] + loc_p[1]) *
-                             bounds[e_divisor]};
-        return (math::fabs(loc_p[0]) <= bounds[e_half_length_0] +
-                                            rel_y * (bounds[e_half_length_1] -
-                                                     bounds[e_half_length_0]) +
-                                            tol and
-                math::fabs(loc_p[1]) <= bounds[e_half_length_2] + tol);
+        const scalar_t rel_y =
+            (bounds[e_half_length_2] + loc_p[1]) * bounds[e_divisor];
+        return (math::fabs(loc_p[0]) <= (bounds[e_half_length_0] +
+                                         rel_y * (bounds[e_half_length_1] -
+                                                  bounds[e_half_length_0]) +
+                                         tol) &&
+                math::fabs(loc_p[1]) <= (bounds[e_half_length_2] + tol));
     }
 
     /// @brief Measure of the shape: Area