--- a/core/include/detray/navigation/intersection/helix_plane_intersector.hpp
+++ b/core/include/detray/navigation/intersection/helix_plane_intersector.hpp
@@ -120,11 +120,9 @@ struct helix_intersector_impl<cartesian2D<algebra_t>, algebra_t> {
         sfi.status = mask.is_inside(sfi.local, tol);
 
         // Compute some additional information if the intersection is valid
-        if (sfi.status == intersection::status::e_inside) {
+        if (sfi.status) {
             sfi.sf_desc = sf_desc;
-            sfi.direction = math::signbit(s)
-                                ? intersection::direction::e_opposite
-                                : intersection::direction::e_along;
+            sfi.direction = !math::signbit(s);
             sfi.volume_link = mask.volume_link();
         }