--- a/core/include/detray/navigation/intersection/ray_intersector.hpp
+++ b/core/include/detray/navigation/intersection/ray_intersector.hpp
@@ -12,6 +12,10 @@
 #include "detray/navigation/intersection/ray_cylinder_portal_intersector.hpp"
 #include "detray/navigation/intersection/ray_line_intersector.hpp"
 #include "detray/navigation/intersection/ray_plane_intersector.hpp"
+#include "detray/navigation/intersection/soa/ray_cylinder_intersector.hpp"
+#include "detray/navigation/intersection/soa/ray_cylinder_portal_intersector.hpp"
+#include "detray/navigation/intersection/soa/ray_line_intersector.hpp"
+#include "detray/navigation/intersection/soa/ray_plane_intersector.hpp"
 
 namespace detray {
 
@@ -20,12 +24,12 @@ namespace detray {
 ///
 /// @note specialized into the concrete intersectors for the different local
 /// geometries in the respective header files
-template <typename frame_t, typename algebra_t>
+template <typename frame_t, typename algebra_t, bool is_soa = false>
 struct ray_intersector_impl {};
 
 template <typename shape_t, typename algebra_t>
 using ray_intersector =
     ray_intersector_impl<typename shape_t::template local_frame_type<algebra_t>,
-                         algebra_t>;
+                         algebra_t, detail::is_soa_v<algebra_t>>;
 
 }  // namespace detray