… core/CMakeLists.txt
… core/include/detray/navigation/intersection/intersection.hpp
… core/include/detray/navigation/intersection/ray_plane_intersector.hpp
… core/include/detray/navigation/intersection/soa/ray_cylinder_intersector.hpp
… core/include/detray/navigation/intersection/soa/ray_cylinder_portal_intersector.hpp
… core/include/detray/navigation/intersection/soa/ray_line_intersector.hpp
… core/include/detray/navigation/intersection/soa/ray_plane_intersector.hpp
… core/include/detray/navigation/intersection_kernel.hpp
… core/include/detray/utils/bounding_volume.hpp
… core/include/detray/utils/quadratic_equation.hpp
… extern/algebra-plugins/CMakeLists.txt
… io/include/detray/io/csv/intersection2D.hpp
… plugins/algebra/CMakeLists.txt
… plugins/algebra/vc_soa/CMakeLists.txt
… plugins/algebra/vc_soa/include/detray/plugins/algebra/vc_soa_definitions.hpp
… tests/benchmarks/cpu/CMakeLists.txt
… tests/benchmarks/cpu/intersect_surfaces.cpp
… tests/benchmarks/cpu/intersectors.cpp
… tests/benchmarks/cpu/masks.cpp
… tests/include/detray/test/utils/detector_scanner.hpp
… tests/include/detray/test/utils/planes_along_direction.hpp
… tests/integration_tests/cpu/propagator/covariance_transport.cpp
… tests/integration_tests/cpu/propagator/jacobian_validation.cpp
… tests/unit_tests/cpu/geometry/masks/annulus2D.cpp
… tests/unit_tests/cpu/geometry/masks/cylinder.cpp
… tests/unit_tests/cpu/geometry/masks/line.cpp
… tests/unit_tests/cpu/geometry/masks/rectangle2D.cpp
… tests/unit_tests/cpu/geometry/masks/ring2D.cpp
… tests/unit_tests/cpu/geometry/masks/single3D.cpp
… tests/unit_tests/cpu/geometry/masks/trapezoid2D.cpp