This adds SOA intersectors that are explicitly vectorized in order to find out what the interfaces towards ```algebra_plugins``` need to be to later also include simd for GPU. It is based on the algebra-plugins abstraction layer introduced in #711 , so that it becomes easier to differentiate between scalar types that are fundamental types and simd types. For the same reason, this also introduces a separate boolean handling, since these are boolean masks in SOA memory layouts. A lot of the glue code that is still needed in this PR should eventually become part of algebra-plugins.

This adds the SOA implementations of the ray-surfaces intersectors, which at this point can completely reuse the coordinates, shapes, mask and intersection types in detray, but need special handling for the intersectors themselves. This is mainly due to the way the control flow is handled at surface inside/outside checks. The goal remains to find a unified implementation at some point.

Benchmarks have been added, that can compare the SOA/Vc implementation to any other AOS  algebra-plugin