@beomki-yeo 2024-03-25
Do we need to use SoA for vectorization? SoA is a great memory layout for GPU cache data transfer, but isn't vectorization still possible without SoA?