@niermann999 2024-03-25
> Do we need to use SoA for vectorization? SoA is a great memory layout for GPU cache data transfer, but isn't vectorization still possible without SoA?

Yes it is possible (the old VC plugin used to do that), but with (interleaved)SoA layout vectorization can be much more efficient, because the SoA length allows to fill the vector registers up completely, whereas in non-SoA you are limited by the dimension of the geometry. So, e.g. 3D or 6D types don't fit very well when you have registers that take e.g. 4 or 8 floats, but with SoA layout you can take as much data as the vectorization capability of the CPU allows.