@niermann999 2024-03-26
Some preliminary results in single precision:
```
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 0.38, 1.07, 0.53
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_INTERSECT_PLANES_AOS                    1106 ms         1104 ms            1
BM_INTERSECT_PLANES_SOA                     349 ms          348 ms            2
BM_INTERSECT_CYLINDERS_AOS                 6999 ms         6987 ms            1
BM_INTERSECT_CYLINDERS_SOA                 1584 ms         1581 ms            1
BM_INTERSECT_CONCETRIC_CYLINDERS_AOS       3144 ms         3138 ms            1
BM_INTERSECT_CONCETRIC_CYLINDERS_SOA        942 ms          940 ms            1
```