@beomki-yeo 2024-04-06
I think it makes sense that Vectorization with SoA is much faster than AoS.
Meanwhile, can we try to improve the performance of AoS even further?

As I understood, we can even try to draw the more potential of CPU vectorization just by tweaking the order of operations, and etc.