@niermann999 2024-04-06
> I think it makes sense that Vectorization with SoA is much faster than AoS. Meanwhile, can we try to improve the performance of AoS even further?
> 
> As I understood, we can even try to draw the more potential of CPU vectorization just by tweaking the order of operations, and etc.

I would rather like to put explicit vectorization back into the old Vc plugin, instead of relying on autovectorization (I tried to do that a little bit when I did the vectorization feasibility study and I found it was a bit fiddly, but that could just be me). That plugin also has the correct memory layout and alignment, tries to minimize elementwise access etc.