@beomki-yeo 2024-04-06
The auto-vectorization with AoS won't be as good as the SoA but I believe we can narrow down the performance gap between two memory layouts. Anyway performance improvement with auto-vectorization is free lunch so worth investigating in the future.
 
I am not inclined in using Vc library for AoS vectorization which is going to introduce another plugin. 
Instead, we can utilize the [Intel advisor](https://www.intel.com/content/www/us/en/docs/advisor/user-guide/2024-1/analyze-vectorization-perspective.html) to improve the cmath plugin