@niermann999 2024-04-06
> The auto-vectorization with AoS won't be as good as the SoA but I believe we can narrow down the performance gap between two memory layouts. Anyway performance improvement with auto-vectorization is free lunch so worth investigating in the future.
> 
> I am not inclined in using Vc library for AoS vectorization which is going to introduce another plugin. Instead, we can utilize the [Intel advisor](https://www.intel.com/content/www/us/en/docs/advisor/user-guide/2024-1/analyze-vectorization-perspective.html) to improve the cmath plugin

True, it comes for free, if would absolutely be great to do! I would expect that it is already autovectorizing the simple stuff, like additions etc, based on the benchmarks I saw in algebra-plugins when switching between debug and Realease build, but I never actually looked at the assembly.
I am not planning to introduce a new plugin though, just to fix the one we already have. I also investigated a third option that showed good performance (not as good as native SoA, but better than AoS), which is using gather operations from AoS to SoA and then doing the computation heavy parts. But for that I needed large, contiguous containers first, which is why I originally refactored the detray geometry back in the beginning of the project.