@niermann999 2024-04-12
OK, this is interesting. After the Vc AoS plugin seemed faster than the cmath plugin in the algebra-plugins benchmarks, I integrated it into detray and lo and behold, the intersector benchmark is 2x slower now
```
Run on (48 X 1796.47 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 4.51, 4.53, 4.73
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_INTERSECT_PLANES_AOS                    2200 ms         2196 ms            1
BM_INTERSECT_PLANES_SOA                     388 ms          387 ms            2
BM_INTERSECT_CYLINDERS_AOS                 7230 ms         7218 ms            1
BM_INTERSECT_CYLINDERS_SOA                 1600 ms         1597 ms            1
BM_INTERSECT_CONCETRIC_CYLINDERS_AOS       6441 ms         6430 ms            1
BM_INTERSECT_CONCETRIC_CYLINDERS_SOA        940 ms          938 ms            1
```
Compiled with gcc ``` cmake -DCMAKE_BUILD_TYPE=Release -DDETRAY_CUSTOM_SCALARTYPE=float ...``` and in the CMakeLists:
```
target_compile_options(detray_algebra_vc_aos
   INTERFACE ${Vc_COMPILE_FLAGS} ${Vc_ARCHITECTURE_FLAGS} )
```
Am I missing something obvious?