@beomki-yeo 2024-04-12
That's weird. Just by adding the following vectorization tag in `algebra-plugins/frontend/array_cmath/CMakeLists.txt` made algebra plugin benchmark faster.

```
target_compile_options(algebra_array_cmath INTERFACE 
                       "-ftree-vectorize" "-march=native" "-mtune=native" "-mavx2")
```                       

Can you put this and see if this also delays the detray benchmark