@beomki-yeo 2024-04-12
I did it by myself It seems faster a little with the vectorization tag (at least it does not drop)

Before

[beomki@beomkilaptop1 detray_oneapi]$ ./bin/detray_benchmark_cuda_array 
2024-04-12T20:13:10+02:00
Running ./bin/detray_benchmark_cuda_array
Run on (12 X 4334.07 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 5.84, 4.66, 3.14
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8      12426379 ns     11092959 ns          460 TracksPropagated=5.76943k/s
CPU unsync propagation/16     17098470 ns     15349246 ns          117 TracksPropagated=16.6783k/s
CPU unsync propagation/32     19738491 ns     18258538 ns           43 TracksPropagated=56.0834k/s
CPU unsync propagation/64     32318694 ns     26718637 ns           48 TracksPropagated=153.301k/s
CPU unsync propagation/128    79425599 ns     65761873 ns           11 TracksPropagated=249.141k/s
CPU unsync propagation/256   257724376 ns    200745644 ns            3 TracksPropagated=326.463k/s
CPU sync propagation/8        12950052 ns     12232556 ns           83 TracksPropagated=5.23194k/s
CPU sync propagation/16       12436607 ns     10466693 ns          176 TracksPropagated=24.4585k/s
CPU sync propagation/32       23656047 ns     21842797 ns           41 TracksPropagated=46.8804k/s
CPU sync propagation/64       31090745 ns     26200694 ns           30 TracksPropagated=156.332k/s
CPU sync propagation/128      79768253 ns     64922494 ns           12 TracksPropagated=252.362k/s
CPU sync propagation/256     253393281 ns    208534798 ns            3 TracksPropagated=314.269k/s
CUDA unsync propagation/8   11238746166 ns   11220001650 ns            1 TracksPropagated=5.7041/s
CUDA unsync propagation/16     2099128 ns      2061449 ns          288 TracksPropagated=124.184k/s
CUDA unsync propagation/32     2555680 ns      2547289 ns          270 TracksPropagated=401.996k/s
CUDA unsync propagation/64     4130158 ns      4111232 ns          167 TracksPropagated=996.295k/s
CUDA unsync propagation/128   12129926 ns     12037245 ns           56 TracksPropagated=1.36111M/s
CUDA unsync propagation/256   51827497 ns     51656218 ns           12 TracksPropagated=1.2687M/s
CUDA sync propagation/8        1410797 ns      1403829 ns          488 TracksPropagated=45.5896k/s
CUDA sync propagation/16       1892639 ns      1883605 ns          353 TracksPropagated=135.91k/s
CUDA sync propagation/32       2229749 ns      2221561 ns          310 TracksPropagated=460.937k/s
CUDA sync propagation/64       3756285 ns      3735708 ns          186 TracksPropagated=1096.45k/s
CUDA sync propagation/128     11405021 ns     11135051 ns           61 TracksPropagated=1.47139M/s
CUDA sync propagation/256     48515256 ns     48335167 ns           13 TracksPropagated=1.35587M/s

After

[beomki@beomkilaptop1 detray_oneapi]$ ./bin/detray_benchmark_cuda_array 
2024-04-12T20:19:33+02:00
Running ./bin/detray_benchmark_cuda_array
Run on (12 X 4191.7 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 4.28, 5.20, 4.04
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8      21195114 ns     18906280 ns           39 TracksPropagated=3.38512k/s
CPU unsync propagation/16     13182799 ns     11985132 ns          119 TracksPropagated=21.3598k/s
CPU unsync propagation/32     22025622 ns     21062614 ns           52 TracksPropagated=48.6169k/s
CPU unsync propagation/64     29093769 ns     24743267 ns           25 TracksPropagated=165.54k/s
CPU unsync propagation/128    76890754 ns     60454725 ns           10 TracksPropagated=271.013k/s
CPU unsync propagation/256   249177129 ns    187278837 ns            4 TracksPropagated=349.938k/s
CPU sync propagation/8        22101698 ns     20831231 ns           31 TracksPropagated=3.07231k/s
CPU sync propagation/16       10701703 ns      9778878 ns          100 TracksPropagated=26.1789k/s
CPU sync propagation/32       23378075 ns     21432812 ns           40 TracksPropagated=47.7772k/s
CPU sync propagation/64       27934986 ns     22655596 ns           30 TracksPropagated=180.794k/s
CPU sync propagation/128      74393729 ns     60665003 ns           10 TracksPropagated=270.073k/s
CPU sync propagation/256     240528702 ns    201143897 ns            4 TracksPropagated=325.816k/s
CUDA unsync propagation/8      1461917 ns      1455542 ns          401 TracksPropagated=43.9699k/s
CUDA unsync propagation/16     2086268 ns      2076711 ns          321 TracksPropagated=123.272k/s
CUDA unsync propagation/32     2625548 ns      2563282 ns          272 TracksPropagated=399.488k/s
CUDA unsync propagation/64     4130233 ns      4112335 ns          169 TracksPropagated=996.028k/s
CUDA unsync propagation/128   12257816 ns     11923428 ns           56 TracksPropagated=1.3741M/s
CUDA unsync propagation/256   52085003 ns     51928086 ns           12 TracksPropagated=1.26205M/s
CUDA sync propagation/8        1366581 ns      1359578 ns          487 TracksPropagated=47.0734k/s
CUDA sync propagation/16       1927394 ns      1917811 ns          362 TracksPropagated=133.486k/s
CUDA sync propagation/32       2270517 ns      2261825 ns          303 TracksPropagated=452.732k/s
CUDA sync propagation/64       3816761 ns      3796786 ns          182 TracksPropagated=1078.81k/s
CUDA sync propagation/128     11076295 ns     11038461 ns           61 TracksPropagated=1.48426M/s
CUDA sync propagation/256     49642115 ns     48542794 ns           12 TracksPropagated=1.35007M/s